Update commercial_products.yml
The P2O5 percentage is wrong. It's not 2% it's 3%. So the values for P and PO4 were also wrong as a result. Here are the calculations I used to arrive at the new P and PO4 values. 0.3 / 2.3 = 0.13043478 which is the new P value. The new PO4 value is 0.13043478 (new P value) * 3.07 = 0.40043478