Update compounds.yml
Did a little research on Plantex. Here's the composition link for it, http://msds.plantprod.com/document/1280 Sodium ferric EDTA is the largest component, 53% by my calculations. That's also by far the least soluble of all the ingredients at 90g/L. So the solubility of Plantex will be based on the solubility of the Sodium ferric EDTA. Since it's only 53% by weight we can increase the solubility accordingly to 170g/L.